+++ app/src/main/java/woowacourse/shopping/view/main/MainActivity.kt
@@ -1,13 +1,17 @@
-package woowacourse.shopping
+package woowacourse.shopping.view.main
 
 import android.os.Bundle
 import androidx.recyclerview.widget.GridLayoutManager
-import woowacourse.shopping.base.BaseActivity
+import woowacourse.shopping.R
 import woowacourse.shopping.data.DummyProducts
 import woowacourse.shopping.databinding.ActivityMainBinding
 import woowacourse.shopping.domain.Product
+import woowacourse.shopping.view.base.BaseActivity
+import woowacourse.shopping.view.detail.ProductDetailActivity
 
-class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main), ProductsEventHandler {
+class MainActivity :
+    BaseActivity<ActivityMainBinding>(R.layout.activity_main),
+    ProductsEventHandler {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding.rvProductList.adapter = ProductsAdapter(DummyProducts.products, this)
