+++ app/src/main/res/layout/item_product.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="handler"
-            type="woowacourse.shopping.ProductsEventHandler" />
+            type="woowacourse.shopping.view.main.ProductsEventHandler" />
 
     </data>
